requirement.txt
fastapi 
uvicorn 
sqlalchemy 
alembic 
psycopg2-binary 
python-dotenv 
passlib[bcrypt]
python-jose[cryptography]